Move output analysis to test folder
